DM-5 Перенес max-alg в tropical
