+++ trunk/core/include/tropical.h
@@ -1,25 +1,71 @@
 #pragma once
 
-#include "max_algebra.h"
 #include <iostream>
 #include <string>
+#include <Eigen/Core>
+
+template<typename T>
+struct MaxAlg {
+	T scalar;
+	MaxAlg() = default;
+	MaxAlg(T scalar) noexcept : scalar(scalar) {}
+	operator T() const noexcept { return scalar; }
+
+	MaxAlg &operator+=(MaxAlg right) noexcept {
+		scalar = std::max(scalar, right.scalar);
+		return *this;
+	}
+
+	friend MaxAlg operator+(MaxAlg left, MaxAlg right) noexcept {
+		left += right;
+		return left;
+	}
+
+	MaxAlg &operator*=(MaxAlg right) noexcept {
+		scalar *= right.scalar;
+		return *this;
+	}
+
+	MaxAlg &operator/=(MaxAlg right) noexcept {
+		scalar /= right.scalar;
+		return *this;
+	}
+
+	friend MaxAlg operator*(MaxAlg left, MaxAlg right) noexcept {
+		left *= right;
+		return left;
+	}
+
+	friend MaxAlg operator/(MaxAlg left, MaxAlg right) noexcept {
+		left /= right;
+		return left;
+	}
+};
+
+
+template<class T>
+double sqrt(const MaxAlg<T> &val) {
+	return ::sqrt(val.scalar);
+}
+
+inline double d(double p, double q) {
+	return p / q;
+}
+
+template<class T, Eigen::Index Rows, Eigen::Index Cols>
+using MaxAlgMatrix = Eigen::Matrix<MaxAlg<T>, Rows, Cols>;
+
+template<class T, Eigen::Index Rows, Eigen::Index Cols>
+using MaxAlgArray = Eigen::Array<MaxAlg<T>, Rows, Cols>;
+
 
 using MaxAlgMatrixXd = MaxAlgMatrix<double, Eigen::Dynamic, Eigen::Dynamic>;
 using MaxAlgVectorXd = MaxAlgMatrix<double, Eigen::Dynamic, 1>;
 using MaxAlgArrayXXd = MaxAlgArray<double, Eigen::Dynamic, Eigen::Dynamic>;
 using MaxAlgArrayXd = MaxAlgArray<double, Eigen::Dynamic, 1>;
 
-template<typename T>
-void PrintLog(const T &t, std::string hint = {}, std::string delim = "\n") {
-	std::cout << hint << delim;
-	std::cout << t << std::endl;
-}
+
 std::ostream &operator<<(std::ostream &stream, const MaxAlgMatrixXd &matrixXd);
-MaxAlgMatrixXd power(const MaxAlgMatrixXd &matrixXd, int n = 1);
+
 MaxAlgMatrixXd kleene_star(const MaxAlgMatrixXd &mat);
 double spectral_radius(const MaxAlgMatrixXd &mat);
-
-
-inline double d(double p, double q) {
-	return p / q;
-}
